refactor: more gen
