--- frb_example/pure_dart/benchmark/src/manual_benchmarks.dart
@@ -1,27 +1,0 @@
-// ignore_for_file: invalid_use_of_internal_member, invalid_use_of_protected_member
-
-import 'dart:math';
-
-import 'benchmark_utils.dart';
-
-// For a list of primes: http://compoasso.free.fr/primelistweb/page/prime/liste_online_en.php
-class ComputePrimeBenchmark extends EnhancedBenchmarkBase {
-  final int number;
-
-  const ComputePrimeBenchmark(this.number, {super.emitter})
-      : super('ComputePrime_Number$number');
-
-  @override
-  void run() {
-    final ans = isPrime(number);
-    if (!ans) throw Exception('unexpected');
-  }
-
-  bool isPrime(int n) {
-    final sqrtN = sqrt(n);
-    for (var i = 2; i <= sqrtN; ++i) {
-      if (n % i == 0) return false;
-    }
-    return true;
-  }
-}
